refactoring random review code
