+++ frontend/src/features/search-random-review/api/index.ts
@@ -2,14 +2,14 @@ import ApiService from '@/shared/api/api.service';
 import { Review } from '@/shared/api/types/review.type';
 import { SearchParamsOption } from 'ky';
 
-export type Critarea = 'ALL' | 'RANKED' | 'UNRANKED' | '';
+export type Criteria = 'ALL' | 'RANKED' | 'UNRANKED' | '';
 
 interface SearchRandomReviewDTO {
   reviews: Review[];
 }
 
 class SearchRandomReviewService extends ApiService {
-  searchRandomReviews(type: Critarea, listId: number) {
+  searchRandomReviews(type: Criteria, listId: number) {
     const searchParams: SearchParamsOption = {
       listId,
       type,
